secure_distribution has wrong type (#462)
If I am following https://cloudinary.com/documentation/cloudinary_sdks#configuration_parameters correctly the type should be string. But it seems that a boolean is acceptable. Changing to string to see if someone can take a look if this is accurate. I should be able to place my cname in here.